dm-migrate: add this hidden command
This allows to generate a migration script that creates linstor
"resources" (nodes, DRBD resources/volumes,...) as they existed in
drbdmanage.

This is highly experimental (some client flags are not even implemented at
this point in time). But it is good to get that in early so that it gets
maintained when the client learns new tricks.

If you are not a developer, there is nothing to see, this breaks your
toys.