fix error-reports show redirect
If you have a string that contains utf8 runes and you print them on
python2, this is fine. But if you pipe/redirect the output, the encoding
is none and then the default ascii encoding is chosen on python2. This
then fails.

It is not that easy to fix as one would assume, using py2/py3 compatible
print() does not simplify the situation.

As most of our output is printed with the help of our Table class, we
are fine, it handles these situations. In most other cases, we control
the output. Error reports are one of the few things where we get utf8
from an "external source". Therefore, fix this where it happens, don't
try to use a general solution. After studying the "general fixes", I
think they cause more problems than they solve.

This closes LINBIT/linstor-server#13