docker: allow specific version tags
This will make our docker images a bit more useful. In lbbuild we can set the
DOCKER_TAG and get both tags, one with the specific version and one with
"latest". If DOCKER_TAG is not set, we tag latest to latest and push latest
twice. All of these are NOPs.