show resource+snapshot name for backups
Showing resource and snapshot name for backups makes it easier
to decipher the origin of the snapshot. This is also important
when viewing backups for restoring via some plugin such as LINSTOR CSI,
which identifies backups via an explicit snapshot name.

Since the ID is mostly used for automation, which is already using an
API client, no need to show it to the user. If desired, it can still be viewed
with a CLI option.