linstor: always enable keep_alive
When the client is idle for some time, the controller closes the
connection. This is especially noticeable for long-running commands such
as "file modify", where the command then fails because it was idle for
too long.

We have a keep_alive flag in the Linstor API package, so just always use
that.