DRBD-Options: Fix help text
Appending "unit_prefix" to "min" and "max" in a "Range" property had a quite
confusing result:

  --sndbuf-size SNDBUF_SIZE
                        Range: [01, 104857601]; Default: 0; Unit: bytes

Actual "min" is 0 (instead of "01") and "max" is 10485760. The two additional "1"
come from the "unit_prefix" (which is "1" in this example) but could be also a
non-number as an other example shows, which makes the intention a bit clearer:

  --resync-rate RESYNC_RATE
                        Range: [1k, 8388608k]; Default: 250; Unit: KiB/s

However, the shown help text of --resync-rate is also wrong. --resync-rate has a
range of 1 byte/sec to 8 MiB/sec. If the user omits any value "250k" (per second)
is taken. If the user enters a value without a unit, "k" is taken as a unit.
This means the following:
* If the property is not specified at all, the default is 250 KiB/s.
* "--resync-rate 100"  results in 100KiB/s.
* "--resync-rate 100k" also results in 100KiB/s.
* "--resync-rate 2M"   results in 2MiB/s.

This patch changes the help text (hopefully) addressing these issues:

  --resync-rate RESYNC_RATE
                        Range: [1, 8388608] in bytes/second; Default value:
                        250; Default unit: k
  --sndbuf-size SNDBUF_SIZE
                        Range: [0, 10485760] in bytes; Default value: 0